Fix to separate module to sweepcache
